feat: Implement comprehensive donation and rewards system with Firebase integration
- Add Firebase and FlutterFire for notifications and analytics
- Implement one-time donation functionality
- Add recurring donation scheduling and management
- Implement round-up donation feature with bank integration
- Add comprehensive rewards system with points and redemption
- Implement store profile and business integration
- Add badges progress tracking system
- Update profile management with transaction history
- Configure flutter_local_notifications for Android and iOS
- Update Android compileSdk to 36 for compatibility
- Enhance UI components across donation and rewards features

+++ lib/features/donation/models/roundup_stats_model.dart
@@ -21,7 +21,7 @@ class RoundupStatsResponse {
 class RoundupStats {
   final double currentRoundupBalance;
   final double todaysRoundupAmount;
-  final int monthlyThreshold;
+  final double monthlyThreshold;
   final double lastTransactionAmount;
   final double roundupPercentage;
   final int daysLeft;
@@ -41,7 +41,7 @@ class RoundupStats {
     return RoundupStats(
       currentRoundupBalance: json['currentRoundupBalance'].toDouble(),
       todaysRoundupAmount: json['todaysRoundupAmount'].toDouble(),
-      monthlyThreshold: json['monthlyThreshold'],
+      monthlyThreshold: json['monthlyThreshold'].toDouble(),
       lastTransactionAmount: json['lastTransactionAmount'].toDouble(),
       roundupPercentage: json['roundupPercentage'].toDouble(),
       daysLeft: json['daysLeft'],
